+++ packages/middleware-content-length/src/middleware-content-length.integ.spec.ts
@@ -29,8 +29,6 @@ describe("middleware-content-length", () => {
         },
       });
 
-      console.log(client.middlewareStack);
-
       await client.createCity({
         name: "MyCity",
         coordinates: {