+++ smithy-typescript-codegen/src/main/java/software/amazon/smithy/typescript/codegen/TypeScriptDependency.java
@@ -113,7 +113,7 @@ public enum TypeScriptDependency implements Dependency {
     AWS_SDK_QUERYSTRING_BUILDER("dependencies", "@smithy/querystring-builder", false),
 
     // Conditionally added when XML parser needs to be used.
-    XML_PARSER("dependencies", "fast-xml-parser", "4.2.5", false),
+    XML_PARSER("dependencies", "fast-xml-parser", "4.4.1", false),
     HTML_ENTITIES("dependencies", "entities", "2.2.0", false),
 
     // Conditionally added when streaming blob response payload exists.
@@ -124,7 +124,6 @@ public enum TypeScriptDependency implements Dependency {
     // Conditionally added when @aws.auth#sigv4 is used
     SIGNATURE_V4("dependencies", "@smithy/signature-v4", false),
 
-    // feat(experimentalIdentityAndAuth): Conditionally added dependencies for `experimentalIdentityAndAuth`.
     // This package should never have a major version, and should only use minor and patch versions in development.
     // Exports are located between @smithy/types and @smithy/core
     @Deprecated EXPERIMENTAL_IDENTITY_AND_AUTH("dependencies", "@smithy/experimental-identity-and-auth", false),