+++ smithy-typescript-codegen/src/main/java/software/amazon/smithy/typescript/codegen/auth/http/integration/SupportHttpApiKeyAuth.java
@@ -23,8 +23,6 @@ import software.amazon.smithy.utils.SmithyInternalApi;
 
 /**
  * Support for @httpApiKeyAuth.
- *
- * This is the experimental behavior for `experimentalIdentityAndAuth`.
  */
 @SmithyInternalApi
 public class SupportHttpApiKeyAuth implements HttpAuthTypeScriptIntegration {
@@ -51,11 +49,11 @@ public class SupportHttpApiKeyAuth implements HttpAuthTypeScriptIntegration {
         .build();
 
     /**
-     * Integration should only be used if `experimentalIdentityAndAuth` flag is true.
+     * Integration should be skipped if the `useLegacyAuth` flag is true.
      */
     @Override
     public boolean matchesSettings(TypeScriptSettings settings) {
-        return settings.getExperimentalIdentityAndAuth();
+        return !settings.useLegacyAuth();
     }
 
     @Override