+++ smithy-typescript-codegen/src/main/java/software/amazon/smithy/typescript/codegen/integration/AddBuiltinPlugins.java
@@ -1,29 +1,33 @@
-package software.amazon.smithy.typescript.ssdk.codegen.test.utils;
+/*
+ * Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+package software.amazon.smithy.typescript.codegen.integration;
 
 import static software.amazon.smithy.typescript.codegen.integration.RuntimeClientPlugin.Convention.HAS_CONFIG;
 import static software.amazon.smithy.typescript.codegen.integration.RuntimeClientPlugin.Convention.HAS_MIDDLEWARE;
 
 import java.util.List;
+import software.amazon.smithy.model.shapes.ServiceShape;
+import software.amazon.smithy.rulesengine.traits.EndpointRuleSetTrait;
 import software.amazon.smithy.typescript.codegen.TypeScriptDependency;
-import software.amazon.smithy.typescript.codegen.integration.RuntimeClientPlugin;
-import software.amazon.smithy.typescript.codegen.integration.TypeScriptIntegration;
-import software.amazon.smithy.utils.ListUtils;
 import software.amazon.smithy.utils.SmithyInternalApi;
 
 /**
- * Adds built-in plugins.
+ * Adds all built-in runtime client plugins to clients.
  */
 @SmithyInternalApi
 public class AddBuiltinPlugins implements TypeScriptIntegration {
-
     @Override
     public List<RuntimeClientPlugin> getClientPlugins() {
         // Note that order is significant because configurations might
         // rely on previously resolved values.
-        return ListUtils.of(
+        return List.of(
             RuntimeClientPlugin.builder()
                 .withConventions(
                     TypeScriptDependency.CONFIG_RESOLVER.dependency, "CustomEndpoints", HAS_CONFIG)
+                .servicePredicate((m, s) -> !isEndpointsV2Service(s))
                 .build(),
             RuntimeClientPlugin.builder()
                 .withConventions(TypeScriptDependency.MIDDLEWARE_RETRY.dependency, "Retry")
@@ -33,5 +37,8 @@ public class AddBuiltinPlugins implements TypeScriptIntegration {
                     HAS_MIDDLEWARE)
                 .build());
     }
+
+    private static boolean isEndpointsV2Service(ServiceShape serviceShape) {
+        return serviceShape.hasTrait(EndpointRuleSetTrait.class);
+    }
 }
-