+++ packages/middleware-endpoint/CHANGELOG.md
@@ -1,5 +1,23 @@
 # Change Log
 
+## 3.1.0
+
+### Minor Changes
+
+- 4a40961: add support for accountId in configValueProvider
+
+### Patch Changes
+
+- @smithy/middleware-serde@3.0.3
+
+## 3.0.5
+
+### Patch Changes
+
+- Updated dependencies [d88521e]
+  - @smithy/shared-ini-file-loader@3.1.4
+  - @smithy/node-config-provider@3.1.4
+
 ## 3.0.4
 
 ### Patch Changes