+++ packages/protocol-http/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Change Log
 
+## 4.1.0
+
+### Minor Changes
+
+- 86862ea: switch to static HttpRequest clone method
+
+## 4.0.4
+
+### Patch Changes
+
+- 796567d: add guidance for HttpRequest cloning
+
 ## 4.0.3
 
 ### Patch Changes