+++ private/util-test/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @smithy/util-test
 
+## 0.2.6
+
+### Patch Changes
+
+- Updated dependencies [86862ea]
+  - @smithy/protocol-http@4.1.0
+
+## 0.2.5
+
+### Patch Changes
+
+- Updated dependencies [796567d]
+  - @smithy/protocol-http@4.0.4
+
 ## 0.2.4
 
 ### Patch Changes